+++ src/Products/PloneMeeting/events.py
@@ -951,8 +951,9 @@ def onAdviceAdded(advice, event):
 
     if not advice.advice_hide_during_redaction:
         # Send mail if relevant
-        sendMailIfRelevant(item, 'adviceEdited', 'creators', isSuffix=True)
-        sendMailIfRelevant(item, 'adviceEditedOwner', 'Owner', isRole=True)
+        item.send_suffixes_and_owner_mail_if_relevant("advice_edited")
+        if item.hasMeeting():
+            item.send_suffixes_and_owner_mail_if_relevant("advice_edited_in_meeting")
 
 
 def onAdviceModified(advice, event):
@@ -993,8 +994,9 @@ def onAdviceModified(advice, event):
 
         if not advice.advice_hide_during_redaction:
             # Send mail if relevant
-            sendMailIfRelevant(item, 'adviceEdited', 'creators', isSuffix=True)
-            sendMailIfRelevant(item, 'adviceEditedOwner', 'Owner', isRole=True)
+            item.send_suffixes_and_owner_mail_if_relevant("advice_edited")
+            if item.hasMeeting():
+                item.send_suffixes_and_owner_mail_if_relevant("advice_edited_in_meeting")
 
 
 def onAdviceEditFinished(advice, event):